+++ Sources/CoreData/Repository/ObjectListRepository.swift
@@ -3,7 +3,6 @@
 // Copyright © 2021 Appwise
 //
 
-import Alamofire
 import CoreData
 
 public protocol ObjectListRepository {
@@ -13,7 +12,7 @@ public protocol ObjectListRepository {
 	var fetchRequest: NSFetchRequest<ObjectType> { get }
 
 	var frc: NSFetchedResultsController<ObjectType> { get }
-	func refresh(then handler: @escaping (Result<[ObjectType]>) -> Void)
+	func refresh(then handler: @escaping (Result<[ObjectType], Error>) -> Void)
 	func findOldItems() -> [ObjectType]
 }
 
@@ -29,7 +28,7 @@ public extension ObjectListRepository {
 		)
 	}
 
-	func refresh(then handler: @escaping (Result<[ObjectType]>) -> Void) {
+	func refresh(then handler: @escaping (Result<[ObjectType], Error>) -> Void) {
 		handler(.cancelled)
 	}